services/tenant-webhook/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
kubernetes==26.1.0
PyJWT==2.8.0
cryptography==42.0.5
Flask-Limiter==3.5.0
redis==5.0.1
